requirements.txt
bitarray>=1.2.2
pefile>=2019.4.18
Pillow>=7.1.2
pyparted>=3.11.6
cli-formatter>=1.2.0
PyPDF2>=1.26.0
